Restructure StackFrame and Interpreter code
There is currently a bug where we enter a new frame to early.
When prepare_frame is called, we enter a new frame. This means
the function will return to the same prepare_frame quad over and
over again, causing an infinite loop. To solve this we need to set
the program counter, of the frame before we enter a new one, to the
quad that is after the function call.