Fix issue which allowed overriding declarations in parent lexical scopes
`blocks.sl` is an example of code that would fail before this fix.
Instead of the main function returning 0 (as expected), 2 was returned.
This both break the mutability of the outer-most `a` variable, and is
not intended to happen. This issue was that we forgot to consider
blocks (`{ ... }`) as scopes for variables when creating new result
registers. The `create_register` function found another variable name
`a` in the function, and thus assumed the same register should be used
for it.

To fix this, we add a list of variables instead (with a new map of
variables being created for each block / function). This list is
traversed in all cases, except when we create a new variable
declaration (as we *want* to create a completely new register in that case.)