Force explicity copy() (or require move) of Value
There will soon come changes where the copy constructor is a bit in the
way when wanting to create a new Value. The data of a Value should never
be shared with any other Value instance. So add an explicit copy()
method instead, so we get compilation error when trying to copy a Value
implicitly.