Add ValuePool and PoolEntry
A value pool is intended to a storage for Value instances. Each Value is
owned by some instance of a ValuePool, and can be exposed to users of a
ValuePool by a PoolEntry. User's of a ValuePool can ask for Value of a
certain type to be created. A ValuePool can either be constant or
dynamic, and is also indicated by a PoolEntry.