Make a copy of mutable value when moving
Only Value instances that are mutable can actually change when they are
already in the ValuePool. We currently have no way of detecting if a
Value would ever be mutated after a given state, which means that we are
required to make a copy of the Value before moving it.