Get the address of member holding string data in CStringWrapper
The ffi lib needs a pointer to where the c-string is held. The current
solution is to hold both a std::string and const char* as member, with
the std::string keeping the data alive and the const char* used as input
for the ffi lib. Currently can't think of a better solution, and the
cost is basically the size of a pointer, so not a huge deal.