Print Vector Value instances
We need to have the ability to resolve PoolEntry instances to Value
instances when printing, since Vector holds a list of PoolEntry
instances. In order to not depend on Context in the 'common' component,
we provide an addition ValueResolver abstraction, which is just an
interface for turning a PoolEntry to a Value. Context implements this
interface.

Also update the expected output for tests that print a Vector, as the
trailing comma in not present anymore.