Represent a Value as a variant
By representing a Value as a "lightweight" (24 bytes on my system)
variant, it allows us to mainly pass the Value around as an value (not a
reference or some wrapper / factory). The 24 byte size currently comes
from the 'double' alignment on Real.

The plan is to keep primitive data completely within a Value instance,
and object data (strings and vectors) will keep a pointer to heap data.