Add specific opcodes for creating and adding to vector
If you specify a vector literal in the code, you get a VEC_CREATE
followed by a VEC_ADD for each literal that is declared. This way the
literal values are stored in the constant pool, but the vector is
created during runtime. This is similar to code generated by GCC for
array literals, with each literal just being pushed to the correct
offset on the stack.