Add FUNC_END opcode
This is a noop opcode that should never be interpreted by the
interpreter. It's main purpose is to be able to bind to queued labels
that we will never jump to. Given the following example:

fn main() -> int {
  if (true) {
    return 1;
  } else {
    return 0;
  }
}

We will generate a jump from after the 'return 1' opcode to the quad
after the if statement. However, in this program, there is no quad after
the if statement. When we we're running the interpreter on non-relocated
quads, this was no problem since we would never ever take perform that
jum, since we always do something before it (lite returning or jumping
somewhere else). But now the relocator fails when we try to relocate a
jump quad to a labeled quad which does not exist. These types of jump
can be optimized away, but until we do, we just emit this label black
hole quad.