Stop taking values as references
If we have a Vector wrapped in a Value and take that as a reference,
we're not getting the reference counter increment on the shared_ptr
holding the underlaying Vector data. This actually causes issues when
using '-O2' optimizations, and causes a segfault. So stop taking Value
instances as reference, as they're intended to be taking as value
anyways.